+++ internal/genaiProxyInternal.go
@@ -2,34 +2,42 @@ package internal
 
 import (
 	"bytes"
+	"context"
 	"encoding/json"
-	"fmt"
+	"github.com/checkmarxDev/ai-proxy/pkg/api/redirect_prompt"
 	"github.com/checkmarxDev/gpt-wrapper/pkg/message"
 	"github.com/checkmarxDev/gpt-wrapper/pkg/models"
+	"google.golang.org/grpc"
+	"google.golang.org/grpc/credentials/insecure"
 	"io"
 	"net/http"
 )
 
-type WrapperImpl struct {
-	apiKey        string
-	endPoint      string
+type WrapperInternalImpl struct {
+	connection    *grpc.ClientConn
+	client        redirect_prompt.AiProxyServiceClient
 	dropLen       int
 	setupMessages []message.Message
 }
 
-func NewWrapperImpl(endPoint, apiKey string, dropLen int) Wrapper {
-	return &WrapperImpl{
-		endPoint: endPoint,
-		apiKey:   apiKey,
-		dropLen:  dropLen,
+func NewWrapperInternalImpl(endPoint string, dropLen int) (Wrapper, error) {
+	connection, err := grpc.NewClient(endPoint, grpc.WithTransportCredentials(insecure.NewCredentials()))
+	if err != nil {
+		return nil, err
 	}
+	client := redirect_prompt.NewAiProxyServiceClient(connection)
+	return &WrapperInternalImpl{
+		connection: connection,
+		client:     client,
+		dropLen:    dropLen,
+	}, nil
 }
 
-func (w *WrapperImpl) SetupCall(messages []message.Message) {
+func (w *WrapperInternalImpl) SetupCall(messages []message.Message) {
 	w.setupMessages = messages
 }
 
-func (w *WrapperImpl) Call(requestBody ChatCompletionRequest) (*ChatCompletionResponse, error) {
+func (w *WrapperInternalImpl) Call(requestBody ChatCompletionRequest) (*ChatCompletionResponse, error) {
 	if w.setupMessages != nil {
 		//true for GPT4
 		if requestBody.Model == models.GPT4 {
@@ -43,43 +51,39 @@ func (w *WrapperImpl) Call(requestBody ChatCompletionRequest) (*ChatCompletionRe
 		}
 	}
 
-	req, err := w.prepareRequest(requestBody)
+	req, err := w.prepareRequest(ChatMetaData{}, requestBody)
 	if err != nil {
 		return nil, err
 	}
 
-	resp, err := http.DefaultClient.Do(req)
+	resp, err := w.client.RedirectPrompt(context.Background(), req)
 	if err != nil {
 		return nil, err
 	}
-	defer func() {
-		_ = resp.Body.Close()
-	}()
-
 	return w.handleGptResponse(requestBody, resp)
 }
 
-func (w *WrapperImpl) prepareRequest(requestBody ChatCompletionRequest) (*http.Request, error) {
+func (w *WrapperInternalImpl) prepareRequest(metaData ChatMetaData, requestBody ChatCompletionRequest) (*redirect_prompt.RedirectPromptRequest, error) {
 	jsonData, err := json.Marshal(requestBody)
 	if err != nil {
 		return nil, err
 	}
-	req, err := http.NewRequest(http.MethodPost, w.endPoint, bytes.NewBuffer(jsonData))
-	if err != nil {
-		return nil, err
+	req := &redirect_prompt.RedirectPromptRequest{
+		Tenant:    metaData.TenantID,
+		RequestId: metaData.RequestID,
+		Origin:    metaData.Origin,
+		Content:   jsonData,
 	}
-	req.Header.Set("Content-Type", "application/json")
-	req.Header.Set("Authorization", fmt.Sprintf("Bearer %s", w.apiKey))
 	return req, nil
 }
 
-func (w *WrapperImpl) handleGptResponse(requestBody ChatCompletionRequest, resp *http.Response) (*ChatCompletionResponse, error) {
+func (w *WrapperInternalImpl) handleGptResponse(requestBody ChatCompletionRequest, resp *redirect_prompt.RedirectPromptResponse) (*ChatCompletionResponse, error) {
 	var err error
-	bodyBytes, err := io.ReadAll(resp.Body)
+	bodyBytes, err := io.ReadAll(bytes.NewBuffer(resp.Content))
 	if err != nil {
 		return nil, err
 	}
-	if resp.StatusCode == http.StatusOK {
+	if resp.GenAiErrorCode == http.StatusOK {
 		var responseBody = new(ChatCompletionResponse)
 		err = json.Unmarshal(bodyBytes, responseBody)
 		if err != nil {
@@ -92,7 +96,7 @@ func (w *WrapperImpl) handleGptResponse(requestBody ChatCompletionRequest, resp 
 	if err != nil {
 		return nil, err
 	}
-	switch resp.StatusCode {
+	switch resp.GenAiErrorCode {
 	case http.StatusBadRequest:
 		if errorResponse.Error.Code == errorCodeMaxTokens {
 			return w.Call(ChatCompletionRequest{
@@ -101,9 +105,9 @@ func (w *WrapperImpl) handleGptResponse(requestBody ChatCompletionRequest, resp 
 			})
 		}
 	}
-	return nil, fromResponse(resp.StatusCode, errorResponse)
+	return nil, fromResponse(int(resp.GenAiErrorCode), errorResponse)
 }
 
-func (w *WrapperImpl) Close() error {
-	return nil
+func (w *WrapperInternalImpl) Close() error {
+	return w.connection.Close()
 }
